fix leak cd to permission denied file
