+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rokupin <rokupin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sserbin <sserbin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/16 00:49:29 by rokupin           #+#    #+#             */
-/*   Updated: 2022/02/06 18:59:39 by sserbin          ###   ########.fr       */
+/*   Updated: 2022/02/07 21:12:21 by sserbin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -177,5 +177,6 @@ char		**export_something(char ***env, char **cmd, t_token *lst);
 void		handle_tilda(char **spl, int i, char **env);
 char		*q_remove(char *str, int len);
 int			get_unquoted_len(char *str);
+void		get_prompt(void);
 
 #endif
