+++ src/exec.c
@@ -6,7 +6,7 @@
 /*   By: sserbin <sserbin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/20 01:24:51 by sserbin           #+#    #+#             */
-/*   Updated: 2022/02/06 18:58:26 by sserbin          ###   ########.fr       */
+/*   Updated: 2022/02/07 21:11:33 by sserbin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
