+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: sserbin <sserbin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/16 01:25:37 by rokupin           #+#    #+#             */
-/*   Updated: 2022/02/06 17:18:13 by sserbin          ###   ########.fr       */
+/*   Updated: 2022/02/07 21:09:54 by sserbin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
